requirements.txt
agno
mistralai
python-dotenv
PyPDF2
python-docx
streamlit
